update interfaces and comments
