+++ packages/teams-js/test/public/marketplace.spec.ts
@@ -1,9 +1,8 @@
 import { errorLibraryNotInitialized } from '../../src/internal/constants';
 import { DOMMessageEvent } from '../../src/internal/interfaces';
-import { marketplace } from '../../src/public';
+import { CartStatus, marketplace } from '../../src/public';
 import { app } from '../../src/public/app';
 import { errorNotSupportedOnPlatform, FrameContexts } from '../../src/public/constants';
-import { CartStatus } from '../../src/public/interfaces';
 import { _minRuntimeConfigToUninitialize } from '../../src/public/runtime';
 import { FramelessPostMocks } from '../framelessPostMocks';
 import { MatcherType, validateExpectedArgumentsInRequest } from '../resultValidation';
